pyproject.toml
[project]
name = "Bayesian-LORA"
version = "0.1.0"
description = "Bayesian inference for Low-Rank Adaptation of Large Language Models"
authors = [
    {name = "Bayesian LoRA Team"}
]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision",
    "numpy",
    "tqdm",
    "pyyaml",
    "scikit-learn",
    "scipy",
    "matplotlib",
    "transformers>=4.30.0",
    "datasets>=2.12.0",
    "tokenizers>=0.13.0",
    "peft>=0.4.0",
    "accelerate>=0.20.0",
    "sentencepiece",
    "protobuf",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["bayesian_lora"]
